feat(omnichannel): implement whatsapp and email triage
This implements the findings from the research report to enable
omnichannel AI triage. It correctly maps WhatsApp and Email events
into the unified agent feed items, staging draft responses that
the owner can securely approve via 1-tap on mobile layouts.

This completes the required implementation.
