+++ src/server/monitoring/dashboards/cost_dashboard.json
